+++ ichnaea/monitor/tasks.py
@@ -1,5 +1,10 @@
+from sqlalchemy import func
+
 from ichnaea.async.config import CELERY_QUEUE_NAMES
-from ichnaea.models import ApiKey
+from ichnaea.models import (
+    ApiKey,
+    OCIDCell,
+)
 from ichnaea.async.task import DatabaseTask
 from ichnaea import util
 from ichnaea.worker import celery
@@ -36,6 +41,27 @@ def monitor_api_key_limits(self):
             value = int(v)
             result[name] = value
             stats_client.gauge('apilimit.' + name, value)
+    except Exception:  # pragma: no cover
+        # Log but ignore the exception
+        self.heka_client.raven('error')
+    return result
+
+
+@celery.task(base=DatabaseTask, bind=True, queue='monitor')
+def monitor_ocid_import(self):
+    result = -1
+    try:
+        now = util.utcnow()
+        stats_client = self.stats_client
+        with self.db_session() as session:
+            q = session.query(func.max(OCIDCell.created))
+            max_created = q.first()[0]
+        if max_created:
+            # diff between now and the value, in milliseconds
+            diff = now - max_created
+            result = (diff.days * 86400 + diff.seconds) * 1000
+
+        stats_client.gauge('table.ocid_cell_age', result)
     except Exception:  # pragma: no cover
         # Log but ignore the exception
         self.heka_client.raven('error')